Do not die when a feed entry has no or invalid link
When a feed item has no link or has a link which no installed app can handle,
the app currently hard-crashes with a fatal
`android.content.ActivityNotFoundException: No Activity found to handle Intent`
exception.

Modified so that when a feed item has no link (which is a completely valid
use case for a RSS feed), the link floating action button is not displayed
at all, and when a link is opened, the `ActivityNotFoundException` is caught
and handled with an error dialog instead of crashing.

Also, unified this behavior across other ext-link opening features in the app
by creating a common `Fragment.openLink` extension function.